web test not launching on windows if shell toolchain path contains whitespace (#389)
The windows launcher code does currently break if the path to the bash binary contains
a whitespace. This is common on windows where programs are stored under `C:\Program Files`.

e.g.

```
:run
C:/Program Files/msys2/usr/bin/bash.exe -c "!run_script!"
```

The path needs to be quoted so that it won't be incorrectly picked up as two separate commands. Resulting in an
exception like:

```
'C:/Program' is not recognized as an internal or external command,
operable program or batch file.
```